--- a/packages/@ember/-internals/glimmer/lib/component.ts
+++ b/packages/@ember/-internals/glimmer/lib/component.ts
@@ -13,7 +13,7 @@ import {
 } from '@ember/-internals/views';
 import { assert, deprecate } from '@ember/debug';
 import { DEBUG } from '@glimmer/env';
-import { DirtyableTag } from '@glimmer/reference';
+import { createTag, dirty } from '@glimmer/reference';
 import { normalizeProperty, SVG_NAMESPACE } from '@glimmer/runtime';
 
 import { RootReference, UPDATE } from './utils/references';
@@ -721,7 +721,7 @@ const Component = CoreView.extend(
     init() {
       this._super(...arguments);
       this[IS_DISPATCHING_ATTRS] = false;
-      this[DIRTY_TAG] = DirtyableTag.create();
+      this[DIRTY_TAG] = createTag();
       this[ROOT_REF] = new RootReference(this);
       this[BOUNDS] = null;
 
@@ -776,7 +776,7 @@ const Component = CoreView.extend(
     },
 
     rerender() {
-      this[DIRTY_TAG].inner.dirty();
+      dirty(this[DIRTY_TAG]);
       this._super();
     },