--- a/packages/@ember/-internals/glimmer/lib/helpers/get.ts
+++ b/packages/@ember/-internals/glimmer/lib/helpers/get.ts
@@ -2,12 +2,12 @@ import { set } from '@ember/-internals/metal';
 import { Opaque } from '@glimmer/interfaces';
 import {
   combine,
-  CONSTANT_TAG,
+  createUpdatableTag,
   isConst,
   PathReference,
   Tag,
-  TagWrapper,
   UpdatableTag,
+  update,
   VersionedPathReference,
 } from '@glimmer/reference';
 import { Arguments, NULL_REFERENCE, VM } from '@glimmer/runtime';
@@ -85,7 +85,7 @@ class GetHelperReference extends CachedReference {
   public pathReference: PathReference<string>;
   public lastPath: string | null;
   public innerReference: VersionedPathReference<Opaque>;
-  public innerTag: TagWrapper<UpdatableTag>;
+  public innerTag: UpdatableTag;
   public tag: Tag;
 
   static create(
@@ -111,7 +111,7 @@ class GetHelperReference extends CachedReference {
     this.lastPath = null;
     this.innerReference = NULL_REFERENCE;
 
-    let innerTag = (this.innerTag = UpdatableTag.create(CONSTANT_TAG));
+    let innerTag = (this.innerTag = createUpdatableTag());
 
     this.tag = combine([sourceReference.tag, pathReference.tag, innerTag]);
   }
@@ -122,7 +122,7 @@ class GetHelperReference extends CachedReference {
 
     if (path !== lastPath) {
       innerReference = referenceFromPath(this.sourceReference, path);
-      innerTag.inner.update(innerReference.tag);
+      update(innerTag, innerReference.tag);
       this.innerReference = innerReference;
       this.lastPath = path;
     }