--- a/packages/@ember/-internals/glimmer/lib/utils/outlet.ts
+++ b/packages/@ember/-internals/glimmer/lib/utils/outlet.ts
@@ -1,6 +1,13 @@
 import { Owner } from '@ember/-internals/owner';
 import { Opaque } from '@glimmer/interfaces';
-import { combine, DirtyableTag, Reference, Tag, VersionedPathReference } from '@glimmer/reference';
+import {
+  combine,
+  createTag,
+  dirty,
+  Reference,
+  Tag,
+  VersionedPathReference,
+} from '@glimmer/reference';
 import { Factory as TemplateFactory, OwnedTemplate } from '../template';
 
 export interface RenderState {
@@ -63,7 +70,7 @@ export interface OutletState {
  * Represents the root outlet.
  */
 export class RootOutletReference implements VersionedPathReference<OutletState> {
-  tag = DirtyableTag.create();
+  tag = createTag();
 
   constructor(public outletState: OutletState) {}
 
@@ -77,7 +84,7 @@ export class RootOutletReference implements VersionedPathReference<OutletState> 
 
   update(state: OutletState) {
     this.outletState.outlets.main = state;
-    this.tag.inner.dirty();
+    dirty(this.tag);
   }
 }