… package.json
… packages/@ember/-internals/glimmer/tests/integration/components/tracked-test.js
… packages/@ember/-internals/metal/lib/computed.ts
… packages/@ember/-internals/metal/lib/observer.ts
… packages/@ember/-internals/metal/lib/tags.ts
… packages/@ember/-internals/metal/tests/alias_test.js
… packages/@ember/-internals/metal/tests/observer_test.js
… packages/@ember/-internals/metal/tests/tracked/classic_classes_test.js
… packages/@ember/-internals/metal/tests/tracked/validation_test.js
… packages/@ember/-internals/runtime/lib/mixins/-proxy.js
… packages/@ember/-internals/runtime/lib/system/array_proxy.js
… packages/@ember/-internals/runtime/tests/system/object/computed_test.js
… packages/@ember/object/compat.ts
… packages/@ember/object/tests/computed/dependent-key-compat-test.js
… packages/internal-test-helpers/lib/ember-dev/assertion.ts
… yarn.lock