@pzuraq 2019-08-04
Yeah, definitely aware of this! I always make sure to do explicit checks in hot paths