@pzuraq 2019-08-08
https://github.com/glimmerjs/glimmer-vm/pull/963 is currently ready for review, pending on @wycats and @krisselden currently. Once they review and approve, we can merge that, cut a release, and rebase/merge this PR.