@pzuraq 2019-08-14
The `assert` issue was being caused by Edge 44.17763.1.0, really weird bug 😕